Rename package auth.provider to auth.dependency
